Moving ALL_FEATURES to ESRestTestCase (and make it private - only usage)
