+++ test/framework/src/main/java/org/elasticsearch/test/rest/TestFeatureService.java
@@ -14,18 +14,4 @@ public interface TestFeatureService {
     boolean clusterHasFeature(String featureId);
 
     Set<String> getAllSupportedFeatures();
-
-    TestFeatureService ALL_FEATURES = new TestFeatureService() {
-        @Override
-        public boolean clusterHasFeature(String featureId) {
-            return true;
-        }
-
-        @Override
-        public Set<String> getAllSupportedFeatures() {
-            throw new UnsupportedOperationException(
-                "Only available to properly initialized TestFeatureService. See ESRestTestCase#createTestFeatureService"
-            );
-        }
-    };
 }
